+++ backend/src/database/models/products.model.js
@@ -13,7 +13,7 @@ const ProductModel = (sequelize, DataTypes) => {
         allowNull: false,
       },
       price: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(7, 2),
         allowNull: false,
       },
       qtyStock: {
